+++ tests/testthat/test_mlr_learners_classif_debug.R
@@ -44,20 +44,17 @@ test_that("NA predictions", {
   expect_equal(is.na(p$response), apply(p$prob, 1, anyMissing))
 })
 
-test_that("early_stopping set is available", {
+test_that("test set is available in $.train method", {
   task = tsk("iris")
-  task$set_row_roles(1:10, roles = "early_stopping")
   learner = lrn("classif.debug", save_tasks = TRUE)
-  learner$train(task)
+  resampling = rsmp("cv", folds = 3)
+  resampling$instantiate(task)
 
-  row_roles = learner$model$task_train$row_roles
-  expect_names(names(row_roles), permutation.of = c("use", "holdout", "early_stopping"))
-  expect_equal(row_roles$early_stopping, seq(10))
-
-  rr = resample(task, learner, rsmp("cv", folds = 3), store_models = TRUE)
+  rr = resample(task, learner, resampling, store_models = TRUE)
 
   walk(seq(rr$iters), function(i) {
-    expect_equal(rr$learners[[i]]$model$task_train$row_roles$early_stopping, seq(10))
+    expect_equal(rr$learners[[i]]$model$task_train$row_roles$use, resampling$train_set(i))
+    expect_equal(rr$learners[[i]]$model$task_train$row_roles$test, resampling$test_set(i))
   })
 })